refactor: 패키지 분리를 롤백한다
