+++ src/main/java/study/Game.java
@@ -1,4 +1,4 @@
-package study.step_3;
+package study;
 
 import java.util.ArrayList;
 import java.util.List;
