+++ src/main/java/study/InputView.java
@@ -1,4 +1,4 @@
-package study.step_3;
+package study;
 
 import java.util.Scanner;
