+++ src/main/java/study/Operator.java
@@ -1,4 +1,4 @@
-package study.step_2;
+package study;
 
 import java.util.function.BiFunction;
 import java.util.stream.Stream;
