+++ src/main/java/study/RacingCarApplication.java
@@ -1,4 +1,4 @@
-package study.step_3;
+package study;
 
 public class RacingCarApplication {
     public static void main(String[] args) {
