+++ src/main/java/study/RandomNumber.java
@@ -1,4 +1,4 @@
-package study.step_3;
+package study;
 
 import java.util.Random;
