+++ src/test/java/study/CarTest.java
@@ -1,7 +1,9 @@
-package study.step_3;
+package study;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.Car;
+import study.RandomNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
