+++ src/test/java/study/InputNumberTest.java
@@ -1,6 +1,7 @@
-package study.step_2;
+package study;
 
 import org.junit.jupiter.api.Test;
+import study.InputNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
